Centralize rpcWorker code
