+++ embedded_demos/jbrowse-react-linear-genome-view-vite/package.json
@@ -12,20 +12,20 @@
   },
   "dependencies": {
     "@esbuild-plugins/node-modules-polyfill": "^0.2.2",
-    "@fontsource/roboto": "^4.5.3",
+    "@fontsource/roboto": "^5.0.3",
     "@jbrowse/core": "^2.0.0",
     "@jbrowse/react-linear-genome-view": "^2.0.0",
-    "@types/node": "^18.11.12",
+    "@types/node": "^20.3.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.26",
     "@types/react-dom": "^18.0.9",
-    "@vitejs/plugin-react": "^3.0.0",
+    "@vitejs/plugin-react": "^4.0.1",
     "buffer": "^6.0.3",
     "stream-browserify": "^3.0.0",
-    "typescript": "^4.9.4",
+    "typescript": "^5.1.6",
     "vite": "^4.0.0"
   }
 }
