+++ packages/product-core/package.json
@@ -46,7 +46,9 @@
     "@mui/icons-material": "^5.0.0",
     "@mui/material": "^5.10.17",
     "copy-to-clipboard": "^3.3.1",
+    "librpc-web-mod": "^1.0.0",
     "react-error-boundary": "^4.0.3",
+    "serialize-error": "^11.0.0",
     "shortid": "^2.2.15"
   },
   "peerDependencies": {
