+++ packages/product-core/src/declare.d.ts
@@ -1,0 +1,1 @@
+declare module 'librpc-web-mod'
