+++ packages/product-core/src/index.ts
@@ -1,3 +1,4 @@
 export * from './RootModel'
 export * from './Session'
 export * from './ui'
+export * from './rpcWorker'
