+++ products/jbrowse-desktop/package.json
@@ -86,7 +86,6 @@
     "electron-is-dev": "^2.0.0",
     "electron-updater": "^6.1.1",
     "electron-window-state": "^5.0.3",
-    "librpc-web-mod": "^1.1.5",
     "mobx": "^6.6.0",
     "mobx-react": "^7.5.0",
     "mobx-state-tree": "^5.0.0",
@@ -98,7 +97,6 @@
     "react-error-boundary": "^4.0.3",
     "rxjs": "^7.0.0",
     "sanitize-filename": "^1.6.3",
-    "serialize-error": "^8.0.0",
     "timeago.js": "^4.0.2",
     "tss-react": "^4.4.1",
     "use-query-params": "^2.0.0",
