+++ products/jbrowse-react-app/package.json
@@ -85,7 +85,6 @@
     "copy-to-clipboard": "^3.3.1",
     "crypto-js": "^3.0.0",
     "file-saver": "^2.0.0",
-    "librpc-web-mod": "^1.1.5",
     "mobx": "^6.6.0",
     "mobx-react": "^7.5.0",
     "mobx-state-tree": "^5.0.0",
@@ -94,7 +93,6 @@
     "prop-types": "^15.0.0",
     "react-error-boundary": "^4.0.0",
     "rxjs": "^7.0.0",
-    "serialize-error": "^8.0.0",
     "shortid": "^2.2.15",
     "tss-react": "^4.4.1",
     "use-query-params": "^2.0.0",
