+++ products/jbrowse-react-linear-genome-view/package.json
@@ -68,13 +68,11 @@
     "@jbrowse/product-core": "^2.6.1",
     "@mui/icons-material": "^5.0.0",
     "@mui/material": "^5.10.17",
-    "librpc-web-mod": "^1.1.5",
     "mobx": "^6.6.0",
     "mobx-react": "^7.5.0",
     "mobx-state-tree": "^5.0.0",
     "prop-types": "^15.0.0",
     "rxjs": "^7.0.0",
-    "serialize-error": "^8.0.0",
     "tss-react": "^4.4.1"
   },
   "peerDependencies": {
