+++ products/jbrowse-react-linear-genome-view/src/rpcWorker.ts
@@ -1,14 +1,7 @@
 /* eslint-disable no-restricted-globals */
 import './workerPolyfill'
-
-// @ts-ignore needs to stay because otherwise fails during build time
-import RpcServer from 'librpc-web-mod'
+import { initializeWorker } from '@jbrowse/product-core'
 import { enableStaticRendering } from 'mobx-react'
-
-import PluginManager from '@jbrowse/core/PluginManager'
-import { remoteAbortRpcHandler } from '@jbrowse/core/rpc/remoteAbortSignals'
-import PluginLoader, { PluginDefinition } from '@jbrowse/core/PluginLoader'
-import { serializeError } from 'serialize-error'
 
 // locals
 import corePlugins from './corePlugins'
@@ -17,89 +10,10 @@ import corePlugins from './corePlugins'
 // worker
 enableStaticRendering(true)
 
-interface WorkerConfiguration {
-  plugins: PluginDefinition[]
-  windowHref: string
-}
-
-// waits for a message from the main thread containing our configuration, which
-// must be sent on boot
-function receiveConfiguration() {
-  const configurationP: Promise<WorkerConfiguration> = new Promise(resolve => {
-    function listener(event: MessageEvent) {
-      if (event.data.message === 'config') {
-        resolve(event.data.config as WorkerConfiguration)
-        removeEventListener('message', listener)
-      }
-    }
-    self.addEventListener('message', listener)
-  })
-  postMessage({ message: 'readyForConfig' })
-  return configurationP
-}
-
-async function getPluginManager() {
-  // Load runtime plugins
-  const config = await receiveConfiguration()
-  const pluginLoader = new PluginLoader(config.plugins, {
-    fetchESM: url => import(/* webpackIgnore:true */ url),
-  })
-  pluginLoader.installGlobalReExports(self)
-  const runtimePlugins = await pluginLoader.load(config.windowHref)
-  const plugins = [...corePlugins.map(p => ({ plugin: p })), ...runtimePlugins]
-  const pluginManager = new PluginManager(plugins.map(P => new P.plugin()))
-  pluginManager.createPluggableElements()
-  pluginManager.configure()
-
-  return pluginManager
-}
-
-interface WrappedFuncArgs {
-  rpcDriverClassName: string
-  channel: string
-  [key: string]: unknown
-}
-
-type RpcFunc = (args: unknown, rpcDriverClassName: string) => unknown
-
-function wrapForRpc(func: RpcFunc) {
-  return (args: WrappedFuncArgs) => {
-    const { channel, rpcDriverClassName } = args
-    return func(
-      {
-        ...args,
-        statusCallback: (message: string) => {
-          // @ts-expect-error
-          self.rpcServer.emit(channel, message)
-        },
-      },
-      rpcDriverClassName,
-    )
-  }
-}
-
-getPluginManager()
-  .then(pluginManager => {
-    const rpcConfig = Object.fromEntries(
-      pluginManager
-        .getRpcElements()
-        .map(entry => [entry.name, wrapForRpc(entry.execute.bind(entry))]),
-    )
-
-    // @ts-expect-error
-    self.rpcServer = new RpcServer.Server({
-      ...rpcConfig,
-      ...remoteAbortRpcHandler(),
-      ping: () => {
-        // the ping method is required by the worker driver for checking the
-        // health of the worker
-      },
-    })
-    postMessage({ message: 'ready' })
-  })
-  .catch(error => {
-    postMessage({ message: 'error', error: serializeError(error) })
-  })
+// eslint-disable-next-line @typescript-eslint/no-floating-promises
+initializeWorker(corePlugins, {
+  fetchESM: url => import(/* webpackIgnore:true */ url),
+})
 
 export default () => {
   /* do nothing */
